archive/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yoto-up"
version = "0.1.17"
description = "Yoto Up - desktop tool for managing Yoto cards and playlists"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.13"
authors = [ { name = "xkjq" } ]
keywords = ["yoto","desktop","flet"]

# Runtime dependencies used by the GUI app. We keep a small common set here
# and expose optional extras for target-specific heavy packages. This allows
# installing only the dependencies needed for the chosen Flet target (desktop
# or web) and avoids pulling native/audio packages into the web build.
dependencies = [
  "httpx>=0.24.0",
  "loguru>=0.7.0",
  "rich>=12.0.0",
  "requests>=2.28.0",
  "pydantic>=1.10",
  "pillow>=9.0.0",
  "beautifulsoup4>=4.9.0",
  "mutagen>=1.45.1",
  "textual>=0.2.0",
  "pynput",
  "numpy>=1.25",
  "matplotlib>=3.7",
  "typer>=0.19.0",
  "rapidfuzz>=2.0.0",
  "ffmpeg-normalize>=1.36.0",
  "ffmpeg-binaries",
  # Optional NLP tools useful for desktop keyword extraction
  "nltk>=3.8.0",

  "platformdirs>=3.0",

]

license = "MIT"
license-files = ["LICENSE"]

classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Topic :: Software Development :: User Interfaces",
]

[project.urls]
Homepage = "https://github.com/xkjq/yoto-up"
Source = "https://github.com/xkjq/yoto-up"
BugTracker = "https://github.com/xkjq/yoto-up/issues"

[project.optional-dependencies]
# Desktop target: full feature set (audio libs, numpy, plotting)
gui = [
  "flet[desktop]>=0.28.3",
  "pydub>=0.25",
  "librosa>=0.10",
  "pyloudnorm>=0.1",
]

# Web target: keep minimal/browser-compatible deps. Avoid native audio and
# heavy scientific packages which are not compatible with the browser runtime.
web = [
  "flet[web]",
]

# additional extras for docs/build or dev
docs = ["mkdocs-material", "pymdown-extensions"]

[tool.flet]
app.module = "yoto_up.gui"


[tool.uv]
dev-dependencies = [
    "flet[all]==0.28.3",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["yoto_up*"]

[project.scripts]
yoto = "yoto_up.yoto:app"
gui = "yoto_up.gui:start_gui"
